+++ CHANGELOG.md
@@ -9,3 +9,10 @@ fix readme
 ## 1.0.2
 
 add example
+
+## 1.1.0
+
+- Enhance the example
+- Add RTL support
+- Add documentation
+- Add bar cap shape customization